+++ src/components/check-availability/index.tsx
@@ -1,10 +1,18 @@
 import { SyntheticEvent, useCallback, useState } from "react";
 import LoadingButton from "../button";
 import DateInput from "../inputs/dateInput";
+import Select from "../inputs/select";
+import AddressAutocompleteInput from "../inputs/addressAutocompleteInout";
 
-export default function CheckAvailability() {
+export default function CheckAvailability({
+  variant = 1,
+}: {
+  variant?: number;
+}) {
   const [checkInDate, setCheckInDate] = useState("");
   const [checkOutDate, setCheckOutDate] = useState("");
+  const [noRooms, setNoRooms] = useState("");
+  const [location, setLocation] = useState("");
 
   const [calculating, setCalculating] = useState(false);
 
@@ -40,6 +48,29 @@ export default function CheckAvailability() {
             placeholder="Check-out Date"
             label="Check-out Date"
           />
+          {!(variant === 1) && (
+            <div className="w-full grid grid-cols-1 gap-5">
+              <Select
+                isRequired={true}
+                value={noRooms}
+                setValue={setNoRooms}
+                id="no-of-rooms"
+              >
+                <option value="" disabled>
+                  Select number of bedroom
+                </option>
+                {Array.from({ length: 8 }, (_, index) => (
+                  <option key={index} value={`${index + 1}`}>
+                    {index + 1}
+                  </option>
+                ))}
+              </Select>
+              <AddressAutocompleteInput
+                value={location}
+                setValue={setLocation}
+              />
+            </div>
+          )}
         </div>
         <LoadingButton
           type="submit"