+++ src/components/filterAndSort/filter.tsx
@@ -1,6 +1,6 @@
 export default function Filter() {
   return (
-    <div className=" flex items-center gap-1 text-sm">
+    <div className=" flex items-center gap-1 text-xs">
       <span className=" text-nowrap">Filter by:</span>
       <div className=" flex items-center gap-2 border p-4 rounded-lg">
         <input type="date" title="start date" /> {" - "}