+++ src/components/status/index.tsx
@@ -9,6 +9,10 @@ export default function Status({ status }: { status: string | undefined }) {
         <span className=" p-1 px-3 bg-green-500/15 text-green-500 rounded-full">
           {status}
         </span>
+      ) : status?.toLocaleLowerCase().includes("confirmed") ? (
+        <span className=" p-1 px-3 bg-green-500/15 text-green-500 rounded-full">
+          {status}
+        </span>
       ) : (
         <span className=" p-1 px-3 bg-yellow-500/15 text-yellow-500 rounded-full">
           {status}